requirements.txt
Django==5.2.5
celery==5.3.4
djangorestframework==3.15.1
